chore: splitting db functions into different files
